Fix membrane terminal errors.
