--- src/vs/workbench/contrib/terminal/browser/terminal.contribution.membrane.ts
@@ -1,23 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (c) Microsoft Corporation. All rights reserved.
- *  Licensed under the MIT License. See License.txt in the project root for license information.
- *--------------------------------------------------------------------------------------------*/
-
-import { ITerminalProfileService } from '../common/terminal.js';
-import { TerminalService } from './terminalService.membrane.js';
-import { InstantiationType, registerSingleton } from '../../../../platform/instantiation/common/extensions.js';
-import { ITerminalEditorService, ITerminalGroupService, ITerminalInstanceService, ITerminalService } from './terminal.js';
-import { ITerminalLogService } from '../../../../platform/terminal/common/terminal.js';
-import { TerminalInstanceService } from './terminalInstanceService.membrane.js';
-import { TerminalEditorService } from './terminalEditorService.js';
-import { TerminalGroupService } from './terminalGroupService.js';
-import { TerminalProfileService } from './terminalProfileService.membrane.js';
-import { TerminalLogService } from '../../../../platform/terminal/common/terminalLogService.js';
-
-// Register services
-registerSingleton(ITerminalLogService, TerminalLogService, InstantiationType.Delayed);
-registerSingleton(ITerminalService, TerminalService, InstantiationType.Delayed);
-registerSingleton(ITerminalEditorService, TerminalEditorService, InstantiationType.Delayed);
-registerSingleton(ITerminalGroupService, TerminalGroupService, InstantiationType.Delayed);
-registerSingleton(ITerminalInstanceService, TerminalInstanceService, InstantiationType.Delayed);
-registerSingleton(ITerminalProfileService, TerminalProfileService, InstantiationType.Delayed);
