+++ src/vs/workbench/contrib/terminal/browser/terminal.contribution.ts
@@ -3,137 +3,21 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 
-import { getFontSnippets } from '../../../../base/browser/fonts.js';
-import { KeyCode, KeyMod } from '../../../../base/common/keyCodes.js';
-import { Schemas } from '../../../../base/common/network.js';
-import { URI } from '../../../../base/common/uri.js';
-import * as nls from '../../../../nls.js';
-import { Extensions as DragAndDropExtensions, IDragAndDropContributionRegistry, IDraggedResourceEditorInput } from '../../../../platform/dnd/browser/dnd.js';
-import { SyncDescriptor } from '../../../../platform/instantiation/common/descriptors.js';
+import { ITerminalProfileService } from '../common/terminal.js';
+import { TerminalService } from './terminalService.js';
 import { InstantiationType, registerSingleton } from '../../../../platform/instantiation/common/extensions.js';
-import { Registry } from '../../../../platform/registry/common/platform.js';
+import { ITerminalEditorService, ITerminalGroupService, ITerminalInstanceService, ITerminalService } from './terminal.js';
 import { ITerminalLogService } from '../../../../platform/terminal/common/terminal.js';
-import { TerminalLogService } from '../../../../platform/terminal/common/terminalLogService.js';
-import { registerTerminalPlatformConfiguration } from '../../../../platform/terminal/common/terminalPlatformConfiguration.js';
-import { EditorPaneDescriptor, IEditorPaneRegistry } from '../../../browser/editor.js';
-import { ViewPaneContainer } from '../../../browser/parts/views/viewPaneContainer.js';
-import { WorkbenchPhase, registerWorkbenchContribution2 } from '../../../common/contributions.js';
-import { EditorExtensions, IEditorFactoryRegistry } from '../../../common/editor.js';
-import { IViewContainersRegistry, IViewsRegistry, Extensions as ViewContainerExtensions, ViewContainerLocation } from '../../../common/views.js';
-import { ITerminalProfileService, TERMINAL_VIEW_ID, TerminalCommandId } from '../common/terminal.js';
-import { TerminalEditingService } from './terminalEditingService.js';
-import { registerColors } from '../common/terminalColorRegistry.js';
-import { registerTerminalConfiguration } from '../common/terminalConfiguration.js';
-import { terminalStrings } from '../common/terminalStrings.js';
-import './media/terminal.css';
-import './media/terminalVoice.css';
-import './media/widgets.css';
-import './media/xterm.css';
-import { RemoteTerminalBackendContribution } from './remoteTerminalBackend.js';
-import { ITerminalConfigurationService, ITerminalEditingService, ITerminalEditorService, ITerminalGroupService, ITerminalInstanceService, ITerminalService, TerminalDataTransfers, terminalEditorId } from './terminal.js';
-import { registerTerminalActions } from './terminalActions.js';
-import { setupTerminalCommands } from './terminalCommands.js';
-import { TerminalConfigurationService } from './terminalConfigurationService.js';
-import { TerminalEditor } from './terminalEditor.js';
-import { TerminalEditorInput } from './terminalEditorInput.js';
-import { TerminalInputSerializer } from './terminalEditorSerializer.js';
+import { TerminalInstanceService } from './terminalInstanceService.js';
 import { TerminalEditorService } from './terminalEditorService.js';
 import { TerminalGroupService } from './terminalGroupService.js';
-import { terminalViewIcon } from './terminalIcons.js';
-import { TerminalInstanceService } from './terminalInstanceService.js';
-import { TerminalMainContribution } from './terminalMainContribution.js';
-import { setupTerminalMenus } from './terminalMenus.js';
 import { TerminalProfileService } from './terminalProfileService.js';
-import { TerminalService } from './terminalService.js';
-import { TerminalTelemetryContribution } from './terminalTelemetry.js';
-import { TerminalViewPane } from './terminalView.js';
+import { TerminalLogService } from '../../../../platform/terminal/common/terminalLogService.js';
 
 // Register services
 registerSingleton(ITerminalLogService, TerminalLogService, InstantiationType.Delayed);
-registerSingleton(ITerminalConfigurationService, TerminalConfigurationService, InstantiationType.Delayed);
 registerSingleton(ITerminalService, TerminalService, InstantiationType.Delayed);
 registerSingleton(ITerminalEditorService, TerminalEditorService, InstantiationType.Delayed);
-registerSingleton(ITerminalEditingService, TerminalEditingService, InstantiationType.Delayed);
 registerSingleton(ITerminalGroupService, TerminalGroupService, InstantiationType.Delayed);
 registerSingleton(ITerminalInstanceService, TerminalInstanceService, InstantiationType.Delayed);
 registerSingleton(ITerminalProfileService, TerminalProfileService, InstantiationType.Delayed);
-
-// Register workbench contributions
-// This contribution blocks startup as it's critical to enable the web embedder window.createTerminal API
-registerWorkbenchContribution2(TerminalMainContribution.ID, TerminalMainContribution, WorkbenchPhase.BlockStartup);
-registerWorkbenchContribution2(RemoteTerminalBackendContribution.ID, RemoteTerminalBackendContribution, WorkbenchPhase.AfterRestored);
-registerWorkbenchContribution2(TerminalTelemetryContribution.ID, TerminalTelemetryContribution, WorkbenchPhase.AfterRestored);
-
-// Register configurations
-registerTerminalPlatformConfiguration();
-registerTerminalConfiguration(getFontSnippets);
-
-// Register editor/dnd contributions
-Registry.as<IEditorFactoryRegistry>(EditorExtensions.EditorFactory).registerEditorSerializer(TerminalEditorInput.ID, TerminalInputSerializer);
-Registry.as<IEditorPaneRegistry>(EditorExtensions.EditorPane).registerEditorPane(
-	EditorPaneDescriptor.create(
-		TerminalEditor,
-		terminalEditorId,
-		terminalStrings.terminal
-	),
-	[
-		new SyncDescriptor(TerminalEditorInput)
-	]
-);
-Registry.as<IDragAndDropContributionRegistry>(DragAndDropExtensions.DragAndDropContribution).register({
-	dataFormatKey: TerminalDataTransfers.Terminals,
-	getEditorInputs(data) {
-		const editors: IDraggedResourceEditorInput[] = [];
-		try {
-			const terminalEditors: string[] = JSON.parse(data);
-			for (const terminalEditor of terminalEditors) {
-				editors.push({ resource: URI.parse(terminalEditor) });
-			}
-		} catch (error) {
-			// Invalid transfer
-		}
-		return editors;
-	},
-	setData(resources, event) {
-		const terminalResources = resources.filter(({ resource }) => resource.scheme === Schemas.vscodeTerminal);
-		if (terminalResources.length) {
-			event.dataTransfer?.setData(TerminalDataTransfers.Terminals, JSON.stringify(terminalResources.map(({ resource }) => resource.toString())));
-		}
-	}
-});
-
-// Register views
-const VIEW_CONTAINER = Registry.as<IViewContainersRegistry>(ViewContainerExtensions.ViewContainersRegistry).registerViewContainer({
-	id: TERMINAL_VIEW_ID,
-	title: nls.localize2('terminal', "Terminal"),
-	icon: terminalViewIcon,
-	ctorDescriptor: new SyncDescriptor(ViewPaneContainer, [TERMINAL_VIEW_ID, { mergeViewWithContainerWhenSingleView: true }]),
-	storageId: TERMINAL_VIEW_ID,
-	hideIfEmpty: true,
-	order: 3,
-}, ViewContainerLocation.Panel, { doNotRegisterOpenCommand: true, isDefault: true });
-Registry.as<IViewsRegistry>(ViewContainerExtensions.ViewsRegistry).registerViews([{
-	id: TERMINAL_VIEW_ID,
-	name: nls.localize2('terminal', "Terminal"),
-	containerIcon: terminalViewIcon,
-	canToggleVisibility: true,
-	canMoveView: true,
-	ctorDescriptor: new SyncDescriptor(TerminalViewPane),
-	openCommandActionDescriptor: {
-		id: TerminalCommandId.Toggle,
-		mnemonicTitle: nls.localize({ key: 'miToggleIntegratedTerminal', comment: ['&& denotes a mnemonic'] }, "&&Terminal"),
-		keybindings: {
-			primary: KeyMod.CtrlCmd | KeyCode.Backquote,
-			mac: { primary: KeyMod.WinCtrl | KeyCode.Backquote }
-		},
-		order: 3
-	}
-}], VIEW_CONTAINER);
-
-registerTerminalActions();
-
-setupTerminalCommands();
-
-setupTerminalMenus();
-
-registerColors();
